+++ lithops/scripts/cli.py
@@ -90,13 +90,9 @@ def clean(config, backend, storage, debug, all, force):
         compute_handler = ServerlessHandler(compute_config, internal_storage)
     elif mode == STANDALONE:
         compute_config = extract_standalone_config(config)
-        if all:
-            backend_name = compute_config['backend']
-            compute_config[backend_name]['delete_master'] = True
-            compute_config[backend_name]['force'] = force
         compute_handler = StandaloneHandler(compute_config)
 
-    compute_handler.clean()
+    compute_handler.clean(delete_master=all, force=force)
 
     # Clean object storage temp dirs
     storage = internal_storage.storage