+++ lithops/standalone/backends/aws_ec2/config.py
@@ -23,7 +23,7 @@ DEFAULT_CONFIG_KEYS = {
     'ssh_username': 'ubuntu',
     'ssh_password': str(uuid.uuid4()),
     'target_ami': 'ami-0a8e758f5e873d1c1',  # ubuntu 20.04
-    'spot_price': None,
+    'request_spot_instances': True,
     'delete_on_dismantle': True,
     'max_workers': 100,
     'worker_processes': 1